newself/back_server/requirements.txt
backports.ssl_match_hostname
werkzeug
json-rpc
pika
pymongo
requests
redis